Apply linting corrections for Python 2.7

+++ src/wheelbin.py
@@ -10,7 +10,6 @@ import hashlib
 import json
 import base64
 
-
 __version__ = "1.0.0"
 __author__ = "Grant Patten <grant@gpatten.com>"
 
@@ -44,7 +43,7 @@ def convert_wheel(whl_file):
         print "Deleting py file: %s" % py_file
         os.remove(py_file)
 
-    for root, dirs, files in os.walk(whl_name):
+    for root, _dirs, files in os.walk(whl_name):
         for f in files:
             if f.endswith(".py"):
                 py_name = os.path.join(root, f)
@@ -69,15 +68,19 @@ def convert_wheel(whl_file):
 
 def rewrite_record(record_path):
     """Rewrite the record file with pyc files instead of py files."""
+
     record_data = []
     whl_path = os.path.abspath(os.path.join(record_path, "..", ".."))
 
+    if not os.path.exists(record_path):
+        return
+
     with open(record_path, "r") as record:
-        for file_dest, hash, length in csv.reader(record):
+        for file_dest, hash_, length in csv.reader(record):
             if file_dest.endswith(".py"):
                 # Do not keep py files, replace with pyc files
                 file_length = 0
-                hash = hashlib.new(HASH_TYPE)
+                hash_ = hashlib.new(HASH_TYPE)
 
                 pyc_file = file_dest + "c"
                 with open(os.path.join(whl_path, pyc_file), "rb") as f:
@@ -87,13 +90,14 @@ def rewrite_record(record_path):
                         if not data:
                             break
 
-                        hash.update(data)
+                        hash_.update(data)
                         file_length += len(data)
 
-                hash_value = "%s=%s" % (HASH_TYPE, base64.urlsafe_b64encode(hash.digest()).rstrip('='))
+                hash_value = base64.urlsafe_b64encode(hash_.digest()).rstrip('=')
+                hash_value = "%s=%s" % (HASH_TYPE, hash_value)
                 record_data.append((pyc_file, hash_value, file_length))
             else:
-                record_data.append((file_dest, hash, length))
+                record_data.append((file_dest, hash_, length))
 
     with open(record_path, "w") as record:
         csv.writer(record, lineterminator='\n').writerows(sorted(set(record_data)))
@@ -101,23 +105,29 @@ def rewrite_record(record_path):
 
 def update_version(dist_info_path):
     """Update the whl to include .version in the name."""
-    with open(os.path.join(dist_info_path, "METADATA"), "r") as f:
-        original_metadata = f.read()
 
-    with open(os.path.join(dist_info_path, "METADATA"), "w") as f:
-        for line in original_metadata.splitlines():
-            if line.startswith("Version: "):
-                f.write(line + ".compiled\n")
-            else:
-                f.write(line + "\n")
+    metapath = os.path.join(dist_info_path, "METADATA")
+    if os.path.exists(metapath):
+        with open(metapath, "r") as f:
+            original_metadata = f.read()
 
-    with open(os.path.join(dist_info_path, "metadata.json"), "r") as f:
-        json_metadata = json.load(f)
+        with open(metapath, "w") as f:
+            for line in original_metadata.splitlines():
+                if line.startswith("Version: "):
+                    f.write(line + ".compiled\n")
+                else:
+                    f.write(line + "\n")
 
-    json_metadata["version"] += ".compiled"
+    metapath = os.path.join(dist_info_path, "metadata.json")
+    if os.path.exists(metapath):
 
-    with open(os.path.join(dist_info_path, "metadata.json"), "w") as f:
-        json.dump(json_metadata, f)
+        with open(metapath, "r") as f:
+            json_metadata = json.load(f)
+
+        json_metadata["version"] += ".compiled"
+
+        with open(metapath, "w") as f:
+            json.dump(json_metadata, f)
 
     # Rename dist-info directory
     new_dist_info_path = dist_info_path[:-10] + ".compiled.dist-info"
@@ -126,6 +136,7 @@ def update_version(dist_info_path):
 
 def rezip_whl(whl_name):
     """Rezip the whl file with the new compiled name."""
+
     new_zip_name = whl_name.split("-")
     new_zip_name[1] += ".compiled"
     new_zip_name = '-'.join(new_zip_name)
@@ -140,6 +151,8 @@ def rezip_whl(whl_name):
 
 
 def main(args=None):
+    """Entry point for wheelbin."""
+
     parser = argparse.ArgumentParser(description='Compile all py files in a wheel')
     parser.add_argument("whl_file", help="Path to whl to convert")
     args = parser.parse_args(args)
@@ -148,4 +161,3 @@ def main(args=None):
 
 if __name__ == "__main__":
     main()
-
